fix dir cleanup and add error checking for invalid username

+++ app.py
@@ -3,6 +3,7 @@ import streamlit as st
 import subprocess
 import boto3
 import shutil
+import requests
 from dotenv import load_dotenv
 
 load_dotenv()
@@ -10,7 +11,7 @@ load_dotenv()
 st.set_page_config(page_title="VSCO Analyzer",
                    page_icon=':thought_balloon:', layout='wide')
 
-# Initialize AWS S3 client
+# AWS S3 client
 s3_client = boto3.client('s3')
 
 with st.container():
@@ -19,48 +20,54 @@ with st.container():
     st.markdown(
         "This app analyzes your VSCO data and generates insights about your VSCO profile, posts, and journals.")
 
+def check_vsco_username_exists(vsco_username):
+    url = f"https://vsco.co/{vsco_username}/gallery"
+    response = requests.head(url)
+    return response.status_code == 200
+
 def main():
     vsco_username = st.text_input("Enter VSCO Username", "")
 
-    # Check if the temp_data directory existsm, if not, create the directory
+    # Check if the temp_data directory exists, if not, create the directory
     if not os.path.exists('./temp_data'):
-        #
         os.makedirs('./temp_data')
 
-    # Check if the app has write permissions for the temp_data directory
+    # Check if the app has write permissions for the temp_data directory, if not use command chmod u+rwx <dir name>
     if not os.access('./temp_data', os.W_OK):
         st.warning("Permission denied: Cannot write to temp_data directory.")
 
     # Button to trigger data retrieval and upload to S3
     if st.button("Analyze"):
         if vsco_username:
-            # Call vsco-scraper CLI commands using subprocess
-            subprocess.run(
-                f"vsco-scraper {vsco_username} --getImages --getCollection --getProfile", shell=True, check=True)
+            if check_vsco_username_exists(vsco_username):
+                subprocess.run(
+                    f"vsco-scraper {vsco_username} --getImages --getCollection --getProfile", shell=True, check=True)
 
-            # Move the downloaded data to the temp_data directory
-            local_path_to_data = f"./{vsco_username}"
-            temp_data_dir = "./temp_data"
-            os.makedirs(temp_data_dir, exist_ok=True)
-            for root, _, files in os.walk(local_path_to_data):
-                for file in files:
-                    shutil.move(os.path.join(root, file), temp_data_dir)
+                # Move the downloaded data to the temp_data directory
+                local_path_to_data = f"./{vsco_username}"
+                temp_data_dir = "./temp_data"
+                os.makedirs(temp_data_dir, exist_ok=True)
+                for root, _, files in os.walk(local_path_to_data):
+                    for file in files:
+                        shutil.move(os.path.join(root, file), temp_data_dir)
 
-            # Upload the data from the temp_data directory to the S3 bucket
-            for file in os.listdir(temp_data_dir):
-                s3_key = f'{vsco_username}/{file}'
-                s3_client.upload_file(os.path.join(
-                    temp_data_dir, file), 'vsco-scrapper', s3_key)
+                # Upload the data from the temp_data directory to the S3 bucket
+                for file in os.listdir(temp_data_dir):
+                    s3_key = f'{vsco_username}/{file}'
+                    s3_client.upload_file(os.path.join(
+                        temp_data_dir, file), 'vsco-scrapper', s3_key)
 
-            # Remove the temp_data directory to avoid cluttering
-            shutil.rmtree(temp_data_dir)
-            shutil.rmtree(local_path_to_data)
+                # Remove the temp_data directory to avoid cluttering
+                shutil.rmtree(temp_data_dir)
+                shutil.rmtree(local_path_to_data)
 
-            # Process the downloaded data and perform visualizations
-            # Your code for data processing and visualization here
+                # Process the downloaded data and perform visualizations
+                # Your code for data processing and visualization here
 
-            # Display visualizations using Streamlit's plotting functions
-            # st.plotly_chart, st.image, etc.
+                # Display visualizations using Streamlit's plotting functions
+                # st.plotly_chart, st.image, etc.
+            else:
+                st.warning("Username not found or doesn't exist. Please make sure to spell it correctly.")
         else:
             st.warning("Please enter a valid VSCO username.")
